#16350 query transformer top limit moved from sql server to the model; now used for teradata too
